Effect for composite binary endpoints

This function calculates different effect measures for binary composite outcomes. In particular, it allows for computing the risk difference, risk ratio and odds ratio.

effectsize_cbe(
  p0_e1,
  p0_e2,
  eff_e1,
  effm_e1,
  eff_e2,
  effm_e2,
  effm_ce = "diff",
  rho
)

Arguments

p0_e1
numeric parameter, probability of occurrence E1 in the control group

p0_e2
numeric parameter, probability of occurrence E2 in the control group

eff_e1
numeric parameter, anticipated effect for the composite component E1

effm_e1
Effect measure used for the event E1 (effm_e1 = "diff" for difference of proportions, effm_e1 = "rr" for risk ratio, effm_e1 = "or" for odds ratio)

eff_e2
numeric parameter, anticipated effect for the composite component E2

effm_e2
Effect measure used for the event E2 (effm_e2 = "diff" for difference of proportions, effm_e2 = "rr" for risk ratio, effm_e2 = "or" for odds ratio)

effm_ce
Effect measure used for the composite endpoint (effm_ce = "diff" for difference of proportions, effm_ce = "rr" for risk ratio, effm_ce = "or" for odds ratio)

rho
numeric parameter, Pearson's correlation between the two events E1 and E2

Value

Returns the effect for the composite binary endpoint and the effects for the composite components

Details

The input parameters represent the probability of the composite components and Pearson's correlation between the two components. Note that Pearson's correlation takes values between two bounds that depend on the probabilities p0_e1 and p0_e2. To calculate the correlation bounds you can use the R functions lower_corr and upper_corr, available in this package.

References

Bofill Roig, M., & Gómez Melis, G. (2019). A new approach for sizing trials with composite binary endpoints using anticipated marginal values and accounting for the correlation between components. Statistics in Medicine, 38(11), 1935–1956. https://doi.org/10.1002/sim.8092